Separated UserService and MemberService. Added functionality to show data in membercards on Team members page.
