+++ Alpha_Mvc/Models/TeamMembersViewModel.cs
@@ -2,7 +2,7 @@
 
 public class TeamMembersViewModel
 {
-    public ICollection<UserModel> Users { get; set; } = [];
+    public IEnumerable<UserModel> Users { get; set; } = [];
 
     public UserModel User { get; set; } = new();
     public CreateMemberFormModel CreateMemberFormModel { get; set; } = new();
