+++ Business/Interfaces/IUserService.cs
@@ -1,0 +1,5 @@
+﻿namespace Business.Interfaces;
+
+public interface IUserService
+{
+}
