+++ Business/Services/UserService.cs
@@ -1,24 +1,17 @@
 ﻿using Business.Dtos;
-using Business.Interfaces;
 using Business.Models;
 using Data.Entities;
 using Data.Interfaces;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.WebUtilities;
-using Microsoft.EntityFrameworkCore;
 
 namespace Business.Services;
 
-public class CreateMemberService(IBaseRepository<MemberUserEntity> repository, IBaseRepository<AddressEntity> addressRepository, IBaseRepository<PictureEntity> pictureRepository, UserManager<MemberUserEntity> userManager) : ICreateMemberService
+public class UserService(IBaseRepository<MemberUserEntity> repository, IBaseRepository<AddressEntity> addressRepository, IBaseRepository<PictureEntity> pictureRepository)
 {
     private readonly IBaseRepository<MemberUserEntity> _repository = repository;
     private readonly IBaseRepository<AddressEntity> _addressRepository = addressRepository;
     private readonly IBaseRepository<PictureEntity> _pictureRepository = pictureRepository;
 
-    private readonly UserManager<MemberUserEntity> _userManager = userManager;
-
-
-    public async Task<MemberModel> AddMember(CreateMemberRegForm form)
+    public async Task<MemberModel> AddUser(CreateMemberRegForm form)
     {
         if (form == null)
         {
@@ -96,43 +89,4 @@ public class CreateMemberService(IBaseRepository<MemberUserEntity> repository, I
         }
     }
 
-    public Task<bool> DeleteMember(MemberModel model)
-    {
-        throw new NotImplementedException();
-    }
-
-    public async Task<IEnumerable<MemberModel>> GetAllMembers()
-    {
-        var entities = await _repository.GetAllAsync();
-        var members = new List<MemberModel>();
-        foreach (var entity in entities)
-        {
-
-            var role = await _userManager.GetRolesAsync(entity);
-
-            var jobTitle = role.FirstOrDefault() ?? "No role assigned";
-
-            members.Add(new MemberModel()
-            {
-                Id = Guid.Parse(entity.Id),
-                FirstName = entity.FirstName,
-                LastName = entity.LastName,
-                JobTitle = jobTitle,
-                PhoneNumber = entity.PhoneNumber,
-                Email = entity.Email,
-                ProfileImage = entity.Picture?.PictureUrl ?? ""
-            });
-        }
-        return members;
-    }
-
-    public Task<MemberModel> GetMember(MemberModel model)
-    {
-        throw new NotImplementedException();
-    }
-
-    public Task<MemberModel> UpdateMember(MemberModel model)
-    {
-        throw new NotImplementedException();
-    }
 }
